api creation

+++ pretrained/AraVec.py
@@ -1,4 +1,5 @@
 import gensim
+import os
 import json
 import urllib.request
 import zipfile
@@ -33,16 +34,24 @@ class AraVec:
 
         link = self.models[model_name]
         filename =  "pretrained/" + link.split('/')[-1]
-        urllib.request.urlretrieve (link, filename)
-        print("Model Downloaded")
+        if not os.path.exists(filename):
+            urllib.request.urlretrieve (link, filename)
+            print("Model Downloaded")
 
-        if unzip == True:
+            if unzip == True:
+                with zipfile.ZipFile(filename, 'r') as zipf:
+                    zipf.extractall("pretrained/")
+                model_file = filename.replace('zip','mdl')
+                print("Model Unzipped ")
+
+        elif not os.path.exists(filename.replace('zip','mdl')):
             with zipfile.ZipFile(filename, 'r') as zipf:
-                zipf.extractall()
+                zipf.extractall("pretrained/")
             model_file = filename.replace('zip','mdl')
             print("Model Unzipped ")
+
         else:
-            model_file = filename
+            model_file = filename.replace('zip','mdl')
         return model_file
